requirements.txt
Flask==1.1.1
Flask-PyMongo==2.3.0
google-auth==1.6.3
oauthlib==3.1.0
python-telegram-bot==12.2.0
requests==2.22.0
requests-oauthlib==1.2.0
gunicorn==19.9.0
google-api-python-client==1.7.11
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
python-dateutil==2.8.0
natasha==0.10.0
flask-login==0.4.1
pytz==2019.3
